refactor(parser): start porting arrow function parsing from tsc
relates #3320
