+++ crates/oxc_parser/examples/parser.rs
@@ -15,7 +15,10 @@ fn main() -> Result<(), String> {
     let source_text = std::fs::read_to_string(path).map_err(|_| format!("Missing '{name}'"))?;
     let allocator = Allocator::default();
     let source_type = SourceType::from_path(path).unwrap();
+    let now = std::time::Instant::now();
     let ret = Parser::new(&allocator, &source_text, source_type).parse();
+    let elapsed_time = now.elapsed();
+    println!("{} seconds.", elapsed_time.as_millis());
 
     println!("AST:");
     println!("{}", serde_json::to_string_pretty(&ret.program).unwrap());
