+++ crates/oxc_parser/src/js/mod.rs
@@ -5,6 +5,7 @@
 mod grammar;
 pub mod list;
 
+mod arrow;
 mod binding;
 mod class;
 pub mod declaration;
@@ -14,3 +15,10 @@ mod module;
 mod object;
 mod operator;
 mod statement;
+
+#[derive(Debug, Clone, Copy)]
+pub enum Tristate {
+    True,
+    False,
+    Maybe,
+}
